Add PC-9M wingman with generic controls
Add a wingman that loads from the main PC-9M model without having to duplicate its properties. Values from the current simulator are dynamically applied to this model in real-time for selected properties. This wingman file will serve as the basis for a more generic platform to load or apply to other types of aircraft without needing to manually copy in property list values.